moved county_cluster from demographic to ccdf
