+++ openfisca_us/variables/demographic/household.py
@@ -152,16 +152,3 @@ class county(Variable):
     entity = Household
     label = u"County"
     definition_period = ETERNITY
-
-
-class county_cluster(Variable):
-    value_type = int
-    entity = Household
-    label = u"County Cluster for CCDF NY"
-    definition_period = ETERNITY
-
-    def formula(household, period, parameters):
-        state_code = household("state_code", period).decode_to_str()
-        county = household("county", period).decode_to_str()
-        cluster_mapping = parameters(period).benefit.CCDF.county_cluster
-        return cluster_mapping[state_code][county]
